update delete db api

+++ src/main/java/com/cloud_computing/mariadb/service/impl/ProjectServiceImpl.java
@@ -6,6 +6,7 @@ import com.cloud_computing.mariadb.entity.User;
 import com.cloud_computing.mariadb.exception.BadRequestException;
 import com.cloud_computing.mariadb.exception.ResourceNotFoundException;
 import com.cloud_computing.mariadb.exception.UnauthorizedException;
+import com.cloud_computing.mariadb.responsitory.DbRepository;
 import com.cloud_computing.mariadb.responsitory.ProjectRepository;
 import com.cloud_computing.mariadb.responsitory.UserRepository;
 import com.cloud_computing.mariadb.service.ProjectService;
@@ -13,6 +14,7 @@ import com.cloud_computing.mariadb.util.SecurityUtils;
 import lombok.AccessLevel;
 import lombok.experimental.FieldDefaults;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
@@ -27,6 +29,8 @@ public class ProjectServiceImpl implements ProjectService {
     UserRepository userRepository;
     @Autowired
     ProjectRepository projectRepository;
+    @Autowired
+    private DbRepository dbRepository;
 
     @Override
     public ProjectDTO createProject(ProjectDTO projectDTO) {
@@ -51,6 +55,8 @@ public class ProjectServiceImpl implements ProjectService {
     @Override
     public void deleteProject(Long projectId) {
         Project project = projectRepository.findByUser_UsernameAndId(SecurityUtils.getUsername(), projectId).orElseThrow(() -> new ResourceNotFoundException("Không tìm thấy project."));
+        if(dbRepository.existsByProject_Id(projectId))
+            throw new AccessDeniedException("Phải xóa hết database trong project trước khi xóa project.");
         projectRepository.delete(project);
     }
